requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
httpx==0.25.1
jinja2==3.1.2
python-multipart==0.0.6
